emacs: block-comment-mode Fixed bug when removing whole words
When removing a long word that has parts inside the block comment
boundry, and parts outside. The block comment postfix did not align
with the rest of the comment. Now fixed by changing the way removed
characters work. It now loops, alternating side, until the diff
between the post-change width and the target width is fixed.