emacs: block-comment-mode Added function that gets the widest text
Added function that goes through all rows in the current block comment
and checks the width of the text in said comment rows. The function
then returns the longest of the text in the block comment rows.

This is used together with the width adjustment functions to adjust
the block comment width to the widest block comment text. Currently,
the width alignment assumes that centering is enabled.