emacs: block-comment-mode: fixed bug when inserting new line
When inserting newline while already in a block comment, the offset
from the block comment start was incorrect. The reason being that, if
the previous line only had whitespaces to the right of point, then a
few whitespaces were copied to the next line. This was fixed bo
checking if the copied string only contains whitespaces