Shell: Fixed key mapping
Mapped keys:
* AltGr -> L_Alt
* Caps_Lock -> Control

Use 'xmodmap' instead of 'setxkbmap' when mapping caps lock to
control. When using 'setxkbmap', the mapping is reset when changing
input language. The bindings set with 'xmodmap' is bound to a specific
layout. Thus, swapping to another language removes the bindings, but
they are restored when swapping back to the language in which they
were set.

Updated todo.org