emacs: Add keybinding for flyspell and changed whitespace-mode conf
Add flyspell-correct and flyspell-correct-popup that display a drop
down menu for correcting misspelled words.
Add keybinding <F8> to correct word at point with flyspell and the
drop down menu.

Change whitespace-mode so that the 'auto-cleanup' feature is not
active in makefiles. The feature removes redundant whitespaces and
convert tabs to whitespaces. This hampers .mk files since tabs are
necessary.