emacs: Fix bug in custom yank behaviour
Previously, the kill ring was restored each time a normal yank was invoked. This caused problems
when a number of yank-pop invokations were followed by a kill command. Then the kill ring had not
been restored when the new kill command was invoked, inserting the killed text in the wrong
position.

Fixed by restoring the kill ring each time a yank-pop was invoked instead.